+++ src/lib/redux/combineReducersLegacy.js
@@ -1,5 +1,5 @@
 import { combineReducers as mergeReducers } from 'redux';
-import { bindStateToLocalStorage } from '../';
+import { bindStateToLocalStorage } from '..';
 
 const combineReducers = (pReducers = {}, pStoreName) => {
     return bindStateToLocalStorage(mergeReducers(pReducers), pStoreName);